% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dfi.R
\name{asDfi}
\alias{asDfi}
\title{asDfi}
\usage{
asDfi(x, str_name)
}
\arguments{
\item{x}{A dataframe.}

\item{str_name}{(optional) The name of the new dfi.}
}
\value{
A dfi with the desired name.
}
\description{
Transform a dataframe in a dfi.
}
