% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_}
\alias{filter_}
\title{filter_}
\usage{
filter_(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Other arguments, see filter_ from dtplyr.}
}
\value{
Filtered dataframe.
}
\description{
Filter rows from a dataframe.
}
