% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by_}
\alias{group_by_}
\title{group_by_}
\usage{
group_by_(x, ...)
}
\arguments{
\item{x}{A dfi.}

\item{...}{Other arguments, see group_by_ from dplyr.}
}
\value{
A dfi containing grouped rows from the one in input.
}
\description{
Group by .
}
