% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_join.R
\name{inner_join}
\alias{inner_join}
\title{inner_join}
\usage{
inner_join(x, y, by, ...)
}
\arguments{
\item{x}{A dfi.}

\item{y}{A dfi.}

\item{by}{Columns to join.}

\item{...}{Other arguments, see inner_join from dplyr.}
}
\value{
Inner join result dfi.
}
\description{
Inner join between two dfi.
}
