% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join.R
\name{left_join}
\alias{left_join}
\title{left_join}
\usage{
left_join(x, y, by, ...)
}
\arguments{
\item{x}{A dfi.}

\item{y}{A dfi.}

\item{by}{Columns to join.}

\item{...}{Other arguments, see left_join from dplyr.}
}
\value{
Left join result dfi.
}
\description{
Left join between two dfi.
}
