% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_}
\alias{mutate_}
\title{mutate_}
\usage{
mutate_(x, ...)
}
\arguments{
\item{x}{A dfi.}

\item{...}{Other arguments, see mutate_ from dplyr.}
}
\value{
A dfi containing mutated columns from the one in input.
}
\description{
Mutate columns from a dfi.
}
