% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_assign.R
\name{myAssign}
\alias{myAssign}
\title{newAssignOps}
\usage{
myAssign(x, y)
}
\arguments{
\item{x}{The name of the new variable.}

\item{y}{The value to be assigned to x.}
}
\value{
Creation of the variable x in the current environment.
}
\description{
Definition of the new assignment operator for dataframes.
}
