% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_}
\alias{rename_}
\title{rename_}
\usage{
rename_(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Arguments for the rename_ function from dtplyr.}
}
\value{
The same dataframe in input but with new column names.
}
\description{
Rename columns of a dataframe.
}
