% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_join.R
\name{right_join}
\alias{right_join}
\title{right_join}
\usage{
right_join(x, y, by, ...)
}
\arguments{
\item{x}{A dfi.}

\item{y}{A dfi.}

\item{by}{Columns to join.}

\item{...}{Other arguments, see right_join from dplyr.}
}
\value{
Right join result dfi.
}
\description{
Right join between two dfi.
}
