% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_}
\alias{select_}
\title{select_}
\usage{
select_(x, ...)
}
\arguments{
\item{x}{A dfi.}

\item{...}{Other arguments, see select_ from dplyr.}
}
\value{
A dfi containing selected columns from the one in input.
}
\description{
Select columns from a dfi.
}
