% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_lineage.R
\name{showLineage}
\alias{showLineage}
\title{showLineage}
\usage{
showLineage(dtf_tree, str_type = "dataframes")
}
\arguments{
\item{dtf_tree}{A dataframe containing the lineage.}

\item{str_type}{"dfi" --> Show the lineage of dataframes;
"columns" --> show the lineage of columns.}
}
\value{
A plot containing the lineage.
}
\description{
Show the data lineage.
}
