% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread_}
\alias{spread_}
\title{spread_}
\usage{
spread_(x, key, value)
}
\arguments{
\item{x}{A dfi.}

\item{key}{Key column.}

\item{value}{Value column.}
}
\value{
A dfi New columns created by the spread function.
}
\description{
Spread one column from a dfi.
}
