% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_}
\alias{summarise_}
\title{summarise_}
\usage{
summarise_(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Arguments of the function summerise_ from dtplyr.}
}
\value{
Transformed dataframe.
}
\description{
Summerise a grouped dataframe.
}
