% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{view}
\usage{
view(x, title)
}
\arguments{
\item{x}{A dfi.}

\item{title}{title for viewer window. Defaults to name of x prefixed by Data:.}
}
\value{
Show the dataframe inside the dfi.
}
\description{
view()
}
