% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_}
\alias{filter_}
\title{filter_}
\usage{
filter_(x, ...)
}
\arguments{
\item{x}{A dfi.}

\item{...}{Other arguments, see filter_ from dplyr.}
}
\value{
A dfi containing selected rows from the one in input.
}
\description{
Filter rows from a dfi.
}
