% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_join.R
\name{full_join}
\alias{full_join}
\title{full_join}
\usage{
full_join(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Other arguments, see full_join from dtplyr.}
}
\value{
Full join result dataframe.
}
\description{
Full join between two dataframes.
}
