% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by_}
\alias{group_by_}
\title{group_by_}
\usage{
group_by_(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Other arguments, see group_by_ from dtplyr.}
}
\value{
Grouped dataframe.
}
\description{
Groupby a dataframe.
}
