% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_join.R
\name{inner_join}
\alias{inner_join}
\title{inner_join}
\usage{
inner_join(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Other arguments, see inner_join from dtplyr.}
}
\value{
Inner join result dataframe.
}
\description{
Inner join between two dataframes.
}
