% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join.R
\name{left_join}
\alias{left_join}
\title{left_join}
\usage{
left_join(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Other arguments, see left_join from dtplyr.}
}
\value{
Left join result dataframe.
}
\description{
Left join between two dataframes.
}
