% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_}
\alias{mutate_}
\title{mutate_}
\usage{
mutate_(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Other arguments, see mutate_ from dtplyr.}
}
\value{
Dataframe with new/modified columns.
}
\description{
Create new columns or overwrite old ones.
}
