% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_}
\alias{rename_}
\title{rename_}
\usage{
rename_(x, .dots)
}
\arguments{
\item{x}{A dfi.}

\item{.dots}{Vector of columns names to rename.}
}
\value{
A dfi containing renamed columns.
}
\description{
Rename columns of a dfi.
}
