% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_join.R
\name{right_join}
\alias{right_join}
\title{right_join}
\usage{
right_join(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Other arguments, see right_join from dtplyr.}
}
\value{
Right join result dataframe.
}
\description{
Right join between two dataframes.
}
