% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_}
\alias{select_}
\title{select_}
\usage{
select_(dtf_data, ...)
}
\arguments{
\item{dtf_data}{A dataframe.}

\item{...}{Arguments for the select_ function from dtplyr.}
}
\value{
The same dataframe in input but with selected columns.
}
\description{
Select a subset of columns from a dataframe.
}
