% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_}
\alias{summarise_}
\title{summarise_}
\usage{
summarise_(x, ...)
}
\arguments{
\item{x}{A dfi.}

\item{...}{Other arguments, see summarise_ from dplyr.}
}
\value{
A dfi containing summarised rows from the one in input.
}
\description{
Summarise rows from a dfi.
}
